CLEANUP: Move TS custom sec from `catlog-wasm` to `notebook-types`.
