+++ packages/notebook-types/Cargo.toml
@@ -18,5 +18,5 @@ serde-wasm-bindgen = "0.6.5"
 serde_json = "1.0.140"
 tsify = { version = "0.5", features = ["js"] }
 ustr = { version = "1.1.0", features = ["serde"] }
-uuid = { version = "1.11", features = ["serde", "v7"] }
+uuid = { version = "1.11", features = ["serde"] }
 wasm-bindgen = "0.2.100"
